+++ WeatherAppTest3/app/src/main/java/nikitin/weatherapp/com/weatherapptest3/View/MainWindowFragment.java
@@ -50,7 +50,7 @@ public class MainWindowFragment extends Fragment {
         return fragment;
     }
 
-    public void updateWeather(int activeCityId) {
+    public void updateWeather(long activeCityId) {
         presenter.getWeatherData(activeCityId);
     }
